requirement.txt
Django==2.2
django-mdeditor==0.1.18
django-suit==2.0a1
et-xmlfile==1.0.1
jdcal==1.4.1
Markdown==3.2.2
openpyxl==3.0.4
pytz==2020.1
sqlparse==0.3.1
xlwt==1.3.0
